+++ src/content/masonry/content-masonry.component.jsx
@@ -37,13 +37,42 @@ const MasonryTile = styled.div`
 `;
 
 class ContentMasonry extends React.PureComponent {
+  constructor(props) {
+    super(props);
+    /* If child is a Card, its onExpand callback will be stored here */
+    this.childCallbacks = {};
+  }
+
+  getChildProps = (childIndex) => {
+    const { children } = this.props;
+    const child = children[childIndex];
+    const props = {
+      parent: this.parent,
+      isLastItem: childIndex === children.length - 1 || children.length === undefined,
+    };
+
+    /*  If child is a Card, we will replace its onExpand callback and store
+     the original callback for a later use */
+    if (child && child.type.displayName === 'ContentCard') {
+      props.onExpand = this.updateLayout;
+      this.childCallbacks[child.props.id] = child.props.onExpand;
+    }
+    return props;
+  };
+
+  updateLayout = (...args) => {
+    this.masonry.initializeMasonry();
+    if (args[0] && typeof this.childCallbacks[args[0]] === 'function') this.childCallbacks[args[0]](args);
+    // I think using forceUpdate is justified here
+    this.forceUpdate();
+  };
+
   render() {
     const {
       children, columnCountExtraLarge, columnCountLarge, columnCountMedium, columnCountSmall,
-      topOffset,
-      className,
-      id,
+      topOffset, className, id,
     } = this.props;
+
     return (
       <PerfectScrollbar>
         <Masonry
@@ -59,10 +88,7 @@ class ContentMasonry extends React.PureComponent {
                 md={columnCountMedium}
                 sm={columnCountSmall}
               >
-                {React.cloneElement(child, {
-                  parent: this.parent,
-                  isLastItem: i === children.length - 1 || children.length === undefined,
-                })}
+                {React.cloneElement(child, this.getChildProps(i))}
               </MasonryTile>
             ))}
           </ReactMasonry>